Removed the nw/tools folder and moved or merged the code in elsewhere
