+++ lib/text/analyse.py
