+++ nw/gui/__init__.py
@@ -29,6 +29,7 @@ from nw.gui.elements.viewdetails import GuiDocViewDetails
 
 # Tools
 from nw.gui.tools.dochighlight import GuiDocHighlighter
+from nw.gui.tools.optionstate import OptionState
 from nw.gui.tools.wordcounter import WordCounter
 
 __all__ = [
@@ -54,5 +55,6 @@ __all__ = [
     "GuiSearchBar",
     "GuiDocViewDetails",
     "GuiDocHighlighter",
+    "OptionState",
     "WordCounter",
 ]
