+++ nw/guimain.py
@@ -45,8 +45,7 @@ from nw.gui import (
     GuiConfigEditor, GuiProjectEditor, GuiItemEditor, GuiProjectOutline,
     GuiSessionLogView, GuiDocMerge, GuiDocSplit, GuiProjectLoad
 )
-from nw.core import NWProject, NWDoc, NWIndex
-from nw.tools import countWords
+from nw.core import NWProject, NWDoc, NWIndex, countWords
 from nw.constants import nwFiles, nwItemType, nwAlert
 
 logger = logging.getLogger(__name__)
