--- nw/tools/__init__.py
@@ -1,15 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from nw.tools.analyse import TextAnalysis
-from nw.tools.legacy import projectMaintenance
-from nw.tools.optionstate import OptionState
-from nw.tools.translate import numberToWord
-from nw.tools.wordcount import countWords
-
-__all__ = [
-    "TextAnalysis",
-    "projectMaintenance",
-    "OptionState",
-    "numberToWord",
-    "countWords",
-]
